requirements.txt
aiohttp==3.9.1
aiogram==3.4.1
gspread==5.12.0
google-auth==2.22.0
python-dotenv==1.0.1
matplotlib==3.8.2
numpy==1.26.4
sentry-sdk==1.40.6
boto3==1.34.106
botocore==1.34.106
openpyxl==3.1.2
SQLAlchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
factory_boy==3.3.0
Faker==19.13.0
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
